Componentize Map Marker
