--- components/listings/show/Map.js
@@ -1,43 +1,0 @@
-import React from 'react'
-import MediaQuery from 'react-responsive'
-
-import MapContainer from 'components/shared/Map'
-
-export default class ListingMap extends React.Component {
-  render() {
-    const {listing} = this.props
-    const {lat, lng} = listing.address
-
-    return (
-      <div>
-        <MediaQuery query="(max-width: 600px)">
-          <MapContainer
-            center={{lat, lng}}
-            lat={lat}
-            lng={lng}
-            width="100vw"
-            height="300px"
-          />
-        </MediaQuery>
-
-        <MediaQuery query="(min-width: 601px)">
-          <MapContainer
-            lat={listing.address.lat}
-            lng={listing.address.lng}
-            width="960px"
-            height="500px"
-          />
-        </MediaQuery>
-
-        <style jsx>{`
-          div {
-            display: flex;
-            height: 400px;
-            justify-content: space-around;
-            margin: 100px 0;
-          }
-        `}</style>
-      </div>
-    )
-  }
-}
