+++ components/listings/show/Map/styles.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components'
+
+export default styled.div`
+  height: 400px;
+  margin: 100px 0;
+  div.this {
+    background: black;
+  }
+`
