+++ components/shared/Map/Marker/styles.js
@@ -1,0 +1,6 @@
+import styled from 'styled-components'
+
+export default styled.div`
+  background: red;
+  color: white;
+`
