+++ components/shared/Map/index.js
@@ -1,8 +1,6 @@
 import {Component} from 'react'
 import GoogleMapReact from 'google-map-react'
 
-import MapMarker from 'components/shared/Map/Marker'
-
 export default class MapContainer extends Component {
   static defaultProps = {
     center: {lat: -22.9608099, lng: -43.2096142},
@@ -10,7 +8,7 @@ export default class MapContainer extends Component {
   }
 
   render() {
-    const {markers, center, zoom} = this.props
+    const {children, center, zoom} = this.props
 
     return (
       <GoogleMapReact
@@ -18,11 +16,7 @@ export default class MapContainer extends Component {
         defaultCenter={center}
         defaultZoom={zoom}
       >
-        {markers.map((marker, i) => {
-          console.log('markerrrr', marker)
-          const {lat, lng, text} = marker
-          return <MapMarker key={i} lat={lat} lng={lng} text={text} />
-        })}
+        {children}
       </GoogleMapReact>
     )
   }
